+++ lambda/package.json
@@ -1,25 +1,22 @@
 {
-  "name": "smart-factory-servicio-tizada",
+  "name": "serverless-with-lambda-layer",
   "version": "0.0.0",
-  "description": "Servicio  de Tizada Async",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [
-    "tizada",
-    "aws",
-    "lamba",
-    "node",
-    "javascript"
-  ],
-  "author": "SmartFactory",
+  "description": "This package demonstrates using @sparticuz/chromium as a devDependency with a layer that contains the binaries",
   "license": "ISC",
+  "author": {
+    "name": "Kyle McNally"
+  },
+  "main": "index.js",
+  "scripts": {
+    "deploy": "test/node_modules/.bin/sls deploy",
+    "test": "test/node_modules/.bin/sls invoke --function chromium-test --log"
+  },
   "dependencies": {
-    "chalk": "^5.3.0",
-    "chrome-aws-lambda": "^10.1.0",
-    "puppeteer": "^23.1.0",
-    "puppeteer-core": "^10.1.0"
+    "puppeteer-core": "^19.6.3",
+    "winston": "^3.14.2"
+  },
+  "devDependencies": {
+    "@sparticuz/chromium": "^126.0.0",
+    "serverless": "^3.27.0"
   }
 }
